--- src/redux/authentication/authSelectors.js
@@ -1,10 +1,0 @@
-const getIfLoggedIn = state => state.authentication.loggedIn;
-
-const getUserName = state => state.authentication.user.name;
-
-const authSelectors = {
-  getIfLoggedIn,
-  getUserName,
-};
-
-export default authSelectors;
