--- src/redux/authentication/index.js
@@ -1,4 +1,0 @@
-export { default as authActions } from './authActions';
-export { default as authReducer } from './authReducer';
-export { default as authOperations } from './authOperations';
-export { default as authSelectors } from './authSelectors';
