+++ src/redux/meals/mealsActions.js
@@ -1,0 +1,13 @@
+import { createAction } from '@reduxjs/toolkit';
+
+const addFoodToMealsRequest = createAction('meals/addFoodToMealsRequest');
+const addFoodToMealsSuccess = createAction('meals/addFoodToMealsSuccess');
+const addFoodToMealsError = createAction('meals/addFoodToMealsError');
+
+const mealsActions = {
+  addFoodToMealsRequest,
+  addFoodToMealsSuccess,
+  addFoodToMealsError,
+};
+
+export default mealsActions;
