+++ src/redux/user/userSelectors.js
@@ -1,0 +1,13 @@
+const getIfLoggedIn = (state) => state.user.loggedIn;
+
+const getUser = (state) => state.user.user;
+
+const getUserId = (state) => state.user.user._id;
+
+const userSelectors = {
+  getIfLoggedIn,
+  getUser,
+  getUserId,
+};
+
+export default userSelectors;
